+++ php_xpass.h
@@ -2,17 +2,13 @@
   +----------------------------------------------------------------------+
   | xpass extension for PHP                                              |
   +----------------------------------------------------------------------+
-  | Copyright (c) The PHP Group                                          |
+  | SPDX-FileCopyrightText: Copyright (c) Remi Collet <remi@php.net>     |
   +----------------------------------------------------------------------+
-  | This source file is subject to version 3.01 of the PHP license,      |
-  | that is bundled with this package in the file LICENSE, and is        |
-  | available through the world-wide-web at the following url:           |
-  | http://www.php.net/license/3_01.txt.                                 |
-  | If you did not receive a copy of the PHP license and are unable to   |
-  | obtain it through the world-wide-web, please send a note to          |
-  | license@php.net so we can mail you a copy immediately.               |
-  +----------------------------------------------------------------------+
-  | Author: Remi Collet <remi@php.net>                                   |
+  | This source file is subject to the Modified BSD License that is      |
+  | bundled with this package in the file LICENSE, and is available      |
+  | through the WWW at <https://opensource.org/license/BSD-3-Clause>.    |
+  |                                                                      |
+  | SPDX-License-Identifier: BSD-3-Clause                                |
   +----------------------------------------------------------------------+
 */